feat: Get method is implemented
I implemented the get method of the agency reviews

+++ Go2Climb.API/Go2Climb.API/Domain/Models/AgencyReview.cs
@@ -5,10 +5,10 @@
         public int Id { get; set; }
         public string Date { get; set; }
         public string Comment { get; set; }
-        public short ProfessionalismScore { get; set; }
-        public short SecurityScore { get; set; }
-        public short QualityScore { get; set; }
-        public short CostScore { get; set; }
+        public double ProfessionalismScore { get; set; }
+        public double SecurityScore { get; set; }
+        public double QualityScore { get; set; }
+        public double CostScore { get; set; }
 
         //Relationships
         public string AgencyId { get; set; }
