+++ Go2Climb.API/Go2Climb.API/Domain/Models/ServiceReview.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Date { get; set; }
         public string Comment { get; set; } 
-        public short Score { get; set; }
+        public double Score { get; set; }
         
         //Relationships
         public string ServiceId { get; set; }
